ZxBot/requirements.txt
discord.py==2.3.2
PyQt5==5.15.9